+++ model/app.go
@@ -1,7 +1,0 @@
-package model
-
-type App struct {
-	AppId             string `json:"app_id"`
-	AppSecret         string `json:"app_secret"`
-	VerificationToken string `json:"verification_token"`
-}