+++ CHANGELOG.md
@@ -1,8 +1,14 @@
-# 0709
+# 2022-07-10
+
+- 支持绑定多个app应用
+  - 提供Register接口进行注册
+  - Webhook增加appId标识
+
+# 2022-07-09
 
 - 支持绑定自定义多维表格账本
 - 引入[feishu-bitable-db](https://github.com/geeklubcn/feishu-bitable-db)处理repo操作
 
-# 0703
+# 2022-07-03
 
 - 飞书记账机器人